enabled multi-volume operation
2 major new functionalities:
- ability to execute against a range of devices, limited by Storage
Owner, VA, VP, Source Project
- ability to further manage a range of devices by providing a name mask.
volume_name is no longer a required argument and now works as a mask, it
matches any volume name that starts with the argument.

There is hidden new functionality to match storage owner exactly, based
on path exposure of device to WWNs (it wasn't really required when
device was pin pointed, but is required if we want dynamic device list
generation).