Added execution symmetry on load_hosts
running load_hosts again on the same data will not fail.

for discovered hosts - it will update all available fields
for registered hosts - (there is not much to update) it will just say
the host is already there

the biggest benefit is that this enables a repetitive mechanism to
facilitate situations when passwords change on production hosts in a
regular and massive way.